extract marker test
